feat: added socket message listener and emitter
When user sends a msg, it now first fetches conversation member's socketId
emit the even 'message' with parameters:
- the message Data
- conversation member's socketId
- conversationiD

Added socket.on('message') in useEffect
When even is listened, it checks if the displayed conv if the conversation the msg has been sent to.
It if is, add the message to the messages state tha contains the displayed messages.
If not, does nothing