fix: add conversation on sidebar on conv creation
Added an optionnal parameter to postMessage function? it contains the Conv Data of the conversation the msg has been sent to
The function is called with this message only when a conversation is created.

Due to useState async,
DisplayedConv isn't set yet when the message is being sent sent.So, setMostRecentConv(displayedConv) sets null
When a conversation is created, its object is passed in parameter to the function.
if the parameter exists, setMostRecentConv sets the given parameter
if not, it means it's not a new conversation, then setMostRecentConv(displayedConv)