feat: get recent conv Files (socket part)
Created Files context and medias context:
Once user fetches recent files, it puts them in sessionStorage. Next time, it won't fetch it to backend but just load from sessionStorage.
Which means if others files are received meanwhile, they wont be shown and fetch wont be triggered.
So when a file is sent in a conversation, it checks if there is a non empty sessionStorage data for this conversation, if there is one, it shift the data to it and add it to the context.When a file is sent in a conversation, socket even adds it to the right sessionStorage with concerned context

Also created few utils functions